+++ CHANGES.md
@@ -7,6 +7,7 @@ Features ✨:
 
 Improvements 🙌:
  - Add System theme option and set as default (#904, #2387)
+ - Store megolm outbound session to improve send time of first message after app launch.
  - Warn user when they are leaving a not public room (#1460)
  - Option to disable emoji keyboard (#2563)